pyproject.toml
[project]
name = "astrbot-plugin-liuli"
version = "0.0.1"
description = "娱乐插件"
authors = [
    {name = "xiayuan",email = "xiayuanovo@163.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
